requirements.txt
langchain>=0.2.12
langchain-community>=0.2.10
langchain-openai>=0.1.22
faiss-cpu>=1.7.4
python-dotenv>=1.0.1
tiktoken>=0.7.0
streamlit>=1.36.0
pandas>=2.2.2
numpy>=1.26.4
pytest>=8.2.0
